requirement.txt
dnspython==2.6.1
pillow==10.2.0
pymongo==4.6.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dotenv==1.0.1
xlrd==2.0.1
